requirements.txt
pandas
scikit-learn
pyyaml
numpy
apache-airflow